[SET-260] Separate status bean for serialization
